test(activerecord): move describeIfPg/describeIfMysql to support so all trees share them
