+++ packages/activerecord/src/adapters/postgresql/test-helper.ts
@@ -1,56 +1,14 @@
-import { describe } from "vitest";
-import pg from "pg";
 import { PostgreSQLAdapter } from "../../connection-adapters/postgresql-adapter.js";
 import { pgDatetimeConfig } from "../../connection-adapters/postgresql/pg-datetime-config.js";
 import { Notifications, squish } from "@blazetrails/activesupport";
 import type { NotificationSubscriber, NotificationEvent } from "@blazetrails/activesupport";
-import { postgresUrl } from "../../support/config.js";
+import { pgAvailable, pgHasHintPlan, pgServerVersion } from "../../support/describe-if-pg.js";
 
-// A *serialization* of the PG sub-settings (PGHOST/PGPORT/PGUSER/PGPASSWORD/
-// PGDATABASE), not an env var of its own: these adapter suites probe the server
-// directly via describeIfPg regardless of ARCONN, so they need a URL to hand
-// the driver, but they no longer source one from the environment.
-export const PG_TEST_URL = postgresUrl();
+// The PG server-reachability gate and the probe behind it are tree-neutral —
+// suites outside this tree gate on them — so they live in support/. Re-exported
+// so this tree's tests keep a single import.
+export { describeIfPg, pgServerVersion, PG_TEST_URL } from "../../support/describe-if-pg.js";
 
-let pgAvailable = false;
-let pgServerVersionNum = 0;
-let pgHasHintPlan = false;
-
-async function checkPg(): Promise<{
-  available: boolean;
-  serverVersionNum: number;
-  hasHintPlan: boolean;
-}> {
-  const client = new pg.Client({ connectionString: PG_TEST_URL });
-  try {
-    await client.connect();
-    const res = await client.query<{ v: string }>(
-      "SELECT current_setting('server_version_num') AS v",
-    );
-    const ext = await client.query<{ count: string }>(
-      "SELECT COUNT(*) AS count FROM pg_available_extensions WHERE name = 'pg_hint_plan'",
-    );
-    return {
-      available: true,
-      serverVersionNum: Number(res.rows[0]?.v ?? 0),
-      hasHintPlan: Number(ext.rows[0]?.count ?? 0) > 0,
-    };
-  } catch {
-    return { available: false, serverVersionNum: 0, hasHintPlan: false };
-  } finally {
-    await client.end().catch(() => {});
-  }
-}
-
-({
-  available: pgAvailable,
-  serverVersionNum: pgServerVersionNum,
-  hasHintPlan: pgHasHintPlan,
-} = await checkPg());
-
-export const describeIfPg = pgAvailable ? describe : (describe.skip as typeof describe);
-/** PG server_version_num at module load (0 when unavailable). */
-export const pgServerVersion = pgServerVersionNum;
 /**
  * Mirrors PostgreSQLAdapter#supports_optimizer_hints? — true when the
  * `pg_hint_plan` extension is installed and available. Rails wraps the whole
@@ -59,7 +17,7 @@ export const pgServerVersion = pgServerVersionNum;
  */
 export const pgSupportsOptimizerHints = pgAvailable && pgHasHintPlan;
 /** Mirrors PostgreSQLAdapter#supportsNativePartitioning — PG 10+ (100000). */
-export const pgSupportsNativePartitioning = pgServerVersionNum >= 100000;
+export const pgSupportsNativePartitioning = pgServerVersion >= 100000;
 
 /** Mirrors Rails' with_postgresql_datetime_type — temporarily changes the adapter's datetimeType. */
 export async function withPostgresqlDatetimeType<T>(
